Session table split with statregy
